Added new environment and fixed assets of ROSIntegrationContent folder
The new environment assets are added in gitignore since its size is too big.
All the content from the /Game/ROSIntegrationContent was deleted. Some of the assets were being used from that folder while others were being used from the folder /Game/Content/ROSIntegrationContent. Now everything is being used in the latter foler, what was being used in the former was moved to the latter and everything that was not being used was deleted. The reason of these two folder is that the project in UE version 4.19 the project used one folder that was not visible when converted to 4.21, so a copy was made inside /Game/Content, but the references were still working in the original folder.